DL-927 Fix data types based on api

+++ xchange-finery-markets/src/main/java/org/knowm/xchange/finerymarkets/dto/trade/DealHistory.java
@@ -59,7 +59,7 @@ public class DealHistory {
 
   private long orderId;
 
-  private int clientOrderId;
+  private long clientOrderId;
 
   private long orderPrice;
 
@@ -75,7 +75,7 @@ public class DealHistory {
 
   private long dealMoment;
 
-  private int dealId;
+  private long dealId;
 
   /**
    * Deal aggressor side
